Adding version constraint on brevo

+++ DependencyInjection/SuluFormExtension.php
@@ -106,7 +106,7 @@ class SuluFormExtension extends Extension implements PrependExtensionInterface
                 ];
             }
 
-            if (!\class_exists(\SendinBlue\Client\Configuration::class)) {
+            if (!\class_exists(\Brevo\Types\Configuration::class)) {
                 $additionalResources[BrevoListSelectController::RESOURCE_KEY] = [
                     'routes' => [
                         'list' => 'sulu_form.get_sendinblue_values',
@@ -260,7 +260,10 @@ class SuluFormExtension extends Extension implements PrependExtensionInterface
         }
 
         if ($config['brevo_api_key']) {
-            if (!\class_exists(\Brevo\Types\Configuration::class)) {
+            if (
+                !\class_exists(\Brevo\Types\Configuration::class)
+                && \version_compare(\Composer\InstalledVersions::getVersion('getbrevo/brevo-php') ?? '0', '4.0', '>=')
+            ) {
                 throw new \LogicException('You need to install the "getbrevo/brevo-php" version ^4.0 to use the Brevo type.');
             }
